Refactor Agepro Calculation Launcher method(s) to its own file
